requirements.txt
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
pytube==12.1.3
requests==2.28.2
